refactor(sdk): Extract `RSAKeyPair` creation to environment-specific files
