+++ packages/sdk/src/encryption/RSAKeyPair.ts
@@ -1,16 +1,7 @@
-/**
- * @todo This file contains code for both browser and Node.js environments. Consider
- * making it environment-specific (using separate files or conditional exports), and
- * remove the following "dom" lib reference when done.
- */
-/// <reference lib="dom" />
-
-import crypto from 'crypto'
-import { promisify } from 'util'
 import { KeyExchangeKeyPair } from './KeyExchangeKeyPair'
 import { AsymmetricEncryptionType } from '@streamr/trackerless-network'
-import { utf8ToBinary, getSubtle } from '@streamr/utils'
-import type { webcrypto } from 'crypto'
+import { utf8ToBinary } from '@streamr/utils'
+import { createRSAKeyPair } from '@/createRSAKeyPair'
 
 /**
  * The length of encrypted data determines the minimum length. In StreamrClient we use RSA
@@ -19,34 +10,6 @@ import type { webcrypto } from 'crypto'
  * https://en.wikipedia.org/wiki/Optimal_asymmetric_encryption_padding
  */
 export const MIN_KEY_LENGTH = 640
-
-function ab2str(...args: any[]): string {
-    // @ts-expect-error Uint8Array parameters
-    return String.fromCharCode.apply(null, new Uint8Array(...args) as unknown as number[])
-}
-
-// shim browser btoa for node
-function btoa(str: string | Uint8Array): string {
-    if (global.btoa) { return global.btoa(str as string) }
-    let buffer
-
-    if (Buffer.isBuffer(str)) {
-        buffer = str
-    } else {
-        buffer = Buffer.from(str.toString(), 'binary')
-    }
-
-    return buffer.toString('base64')
-}
-
-async function exportCryptoKey(key: webcrypto.CryptoKey, { isPrivate = false } = {}): Promise<string> {
-    const keyType = isPrivate ? 'pkcs8' : 'spki'
-    const exported = await getSubtle().exportKey(keyType, key)
-    const exportedAsString = ab2str(exported)
-    const exportedAsBase64 = btoa(exportedAsString)
-    const TYPE = isPrivate ? 'PRIVATE' : 'PUBLIC'
-    return `-----BEGIN ${TYPE} KEY-----\n${exportedAsBase64}\n-----END ${TYPE} KEY-----\n`
-}
 
 export class RSAKeyPair implements KeyExchangeKeyPair {
     // the keys are in PEM format
@@ -74,45 +37,7 @@ export class RSAKeyPair implements KeyExchangeKeyPair {
     }
 
     static async create(keyLength: number): Promise<RSAKeyPair> {
-        return (typeof window !== 'undefined')
-            ? RSAKeyPair.create_browserEnvironment(keyLength)
-            : RSAKeyPair.create_serverEnvironment(keyLength)
-    }
-
-    private static async create_serverEnvironment(keyLength: number): Promise<RSAKeyPair> {
-        // promisify here to work around browser/server packaging
-        const generateKeyPair = promisify(crypto.generateKeyPair)
-        const { publicKey, privateKey } = await generateKeyPair('rsa', {
-            modulusLength: keyLength,
-            publicKeyEncoding: {
-                type: 'spki',
-                format: 'pem',
-            },
-            privateKeyEncoding: {
-                type: 'pkcs8',
-                format: 'pem',
-            },
-        })
-
+        const { privateKey, publicKey } = await createRSAKeyPair(keyLength)
         return new RSAKeyPair(privateKey, publicKey)
-    }
-
-    private static async create_browserEnvironment(keyLength: number): Promise<RSAKeyPair> {
-        const { publicKey, privateKey } = await getSubtle().generateKey({
-            name: 'RSA-OAEP',
-            modulusLength: keyLength,
-            publicExponent: new Uint8Array([1, 0, 1]), // 65537
-            hash: 'SHA-256'
-        }, true, ['encrypt', 'decrypt'])
-
-        const [exportedPrivate, exportedPublic] = await Promise.all([
-            exportCryptoKey(privateKey, {
-                isPrivate: true,
-            }),
-            exportCryptoKey(publicKey, {
-                isPrivate: false,
-            })
-        ])
-        return new RSAKeyPair(exportedPrivate, exportedPublic)
     }
 }
