+++ packages/sdk/src/encryption/types.ts
@@ -1,0 +1,4 @@
+export interface PemKeyPair {
+    privateKey: string
+    publicKey: string
+}
